% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{beattrack}
\alias{beattrack}
\title{beat tracking}
\usage{
beattrack(w)
}
\arguments{
\item{w}{a Wave object}
}
\value{
a list. beatpos is a vector of beat positions, boundary is a vector of part boundaries, flux is the spectral flux, localperiod is a vector of local fundamental period, frames is the total frame length
}
\description{
beat tracking
}
