% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{detect.peaks2}
\alias{detect.peaks2}
\title{\code{detect.peaks2}: peak picking}
\usage{
detect.peaks2(values, thres, pwidth, type = "min")
}
\arguments{
\item{values}{a vector from which peaks are found}

\item{thres}{detection threshold}

\item{pwidth}{minimum width that should be kept between peaks}

\item{type}{"min" or "max"}
}
\value{
a vector of positions of peaks
}
\description{
The function \code{detect.peaks2} chooses the local minimum or maximum peaks from the input vector,
and returns the positions of the peaks.
}
