% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{estimatePeriod}
\alias{estimatePeriod}
\title{Estimate fundamental period and BPM}
\usage{
estimatePeriod(nv, bpmlim = c(60, 180), frameshift = 0.01)
}
\arguments{
\item{nv}{input signal (spectral flux is assumed)}

\item{bpmlim}{two-element vector, lower and upper bound of BPM}

\item{frameshift}{The frame period in second (default 0.01)}
}
\value{
list of the BPM and fundamental period.
}
\description{
\code{extimatePeriod} estimates the fundamental period and BPM from the input vector.
}
