% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{finerPeriodAnalysis}
\alias{finerPeriodAnalysis}
\title{Detailed period analysis
\code{finderPeriodAnalysis} analyzes the input signal (spectral flux) in detail.}
\usage{
finerPeriodAnalysis(flux, period, range = 5, prec = 0.01)
}
\arguments{
\item{flux}{the input signal (usually spectral flux)}

\item{period}{an approximate fundamental period (integer)}

\item{range}{search range of fundamental period relative to \code{period}}

\item{prec}{precision of search}
}
\value{
a list of two elements: period is the fundamental period, pos is a vector of beat positions
}
\description{
Detailed period analysis
\code{finderPeriodAnalysis} analyzes the input signal (spectral flux) in detail.
}
