% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{generateBeep}
\alias{generateBeep}
\title{Superimpose beep sound at the beat position
\code{generateBeep} substitutes the right channel of the wave object with beep sounds at the beat positions}
\usage{
generateBeep(org_aud, beatpos, partpos, beeplength = 5, beepamp = 5000)
}
\arguments{
\item{org_aud}{an Wave object}

\item{beatpos}{positions of beat}

\item{partpos}{positions of part boundary}

\item{beeplength}{length of a beep in frames}

\item{beepamp}{amplitude of a beep}
}
\value{
a wave object
}
\description{
Superimpose beep sound at the beat position
\code{generateBeep} substitutes the right channel of the wave object with beep sounds at the beat positions
}
