% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{h_analysis}
\alias{h_analysis}
\title{h_analysis:segmentation of the signal}
\usage{
h_analysis(feature, peaks, hwidth = 500, firstskip = 500)
}
\arguments{
\item{feature}{feature vectors of the song (MFCC), each row is a feature vector.}

\item{peaks}{vector of boundary candidates}

\item{hwidth}{window width (in frame)}

\item{firstskip}{number of frames to skip the first part to be segmented}
}
\value{
a list of two elements: segs is a vector of boundaries, vals is a vector of boundary score
}
\description{
h_analysis:segmentation of the signal
}
