% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{hearst}
\alias{hearst}
\title{Calculate the boundary score using Hearst method
\code{hearst} calculates segmentation score based on Hearst method.}
\usage{
hearst(feature, p, width)
}
\arguments{
\item{feature}{Feature vector (usually MFCC), each row is a feature vector.}

\item{p}{a boundary}

\item{width}{window width in frames}
}
\value{
score of `boundaryness'
}
\description{
Calculate the boundary score using Hearst method
\code{hearst} calculates segmentation score based on Hearst method.
}
