% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{spectralFlux}
\alias{spectralFlux}
\title{Calculate spectral flux}
\usage{
spectralFlux(w, frameshift = 0.01, freq.range = NULL)
}
\arguments{
\item{w}{an Wave object}

\item{frameshift}{The frame shift in second (default 0.01)}

\item{freq.range}{Frequrncy range on which spectral flux is calculated. (NULL means using the entire range) For example, freq.range=c(0,8000) means "use 0 to 8000Hz for calculation"}
}
\value{
a vector of spectral flux
}
\description{
\code{spectfalFlux} calculates the spectral flux from the wave object.
}
