% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat_tracking.R
\name{spectralFlux}
\alias{spectralFlux}
\title{Calculate spectral flux}
\usage{
spectralFlux(w, frameshift = 0.01)
}
\arguments{
\item{w}{an Wave object}

\item{frameshift}{The frame shift in second (default 0.01)}
}
\value{
a vector of spectral flux
}
\description{
\code{spectfalFlux} calculates the spectral flux from the wave object.
}
